pyproject.toml
[project]
name = "frontdesk-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6.0",
    "fastapi>=0.115.6",
    "langchain>=0.3.13",
    "langchain-community>=0.3.13",
    "langchain-nomic>=0.1.4",
    "langchain-openai>=0.2.14",
    "langchainhub>=0.1.21",
    "langgraph>=0.2.60",
    "ngrok>=1.4.0",
    "ollama>=0.4.5",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tavily-python>=0.5.0",
    "tiktoken>=0.8.0",
    "uvicorn>=0.34.0",
]
